ai-providers-config/requirements.txt
# Python dependencies for AI client

python-dotenv==1.0.0          # Environment variable management

# Gemini
google-generativeai==0.3.0

# OpenAI
openai==1.3.0

# Anthropic (Claude)
anthropic==0.7.0

# Mistral
mistralai==0.0.7

# Groq (uses OpenAI SDK)

# AWS Bedrock (optional)
# boto3>=1.26.0

# Cohere
cohere==4.13.0

# Hugging Face
huggingface-hub==0.16.0
requests>=2.31.0

# Replicate
replicate==0.14.0

# Together AI (uses OpenAI SDK)

# Development tools
requests==2.31.0
